Water Map Update
Made the walls (they kill you, no bugs I think) and the mechanism by which the player can stop (a new blueprint called resting spot). I think, besides changing the color of the wall to make sure it's clear they are deadly, this is everything I need to design the puzzle.